Update simplesoapy.py
Changed timeout fudge factor from 0.1 seconds to 2 seconds this extra delay is now needed because because of the change upstream from blocking on first buffer to unblocked. Which adds the initial setup delay of the first buffer. Unfortunately this extra delay on the timeout is only needed for the very first buffer.